Final Updates and modifications
This is the final update for the Ride-Share Simulation. In this update, we have modified the car.py, rider.py, graph.py, and simulation.py. We also renamed the map.csv to legacy_map.csv and created the new city_map.csv. We also included the simulation_summary.png file, which is the result of the simulation.py being run and having the results saved as a visualization in a png file.